+++ packages/e2e-test-app-fabric/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  extends: ['@rnw-scripts'],
+  parserOptions: { tsconfigRootDir: __dirname },
+  globals: {
+    expect: 'readonly',
+    fail: 'readonly',
+    AutomationClient: 'readonly',
+  },
+};