+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionHwndHost.cpp
@@ -14,6 +14,7 @@
 #include "CompositionContextHelper.h"
 #include "ReactNativeHost.h"
 
+#include "CompositionRootAutomationProvider.h"
 #include "CompositionRootView.h"
 
 WINUSERAPI UINT WINAPI GetDpiForWindow(_In_ HWND hwnd);
@@ -154,7 +155,12 @@ winrt::Windows::UI::Composition::Compositor CompositionHwndHost::Compositor() co
 
 IInspectable CompositionHwndHost::UiaProvider() noexcept {
   auto compRootView = winrt::get_self<implementation::CompositionRootView>(m_compRootView);
-  return compRootView->GetUiaProvider(reinterpret_cast<uint64_t>(m_hwnd));
+  auto provider = compRootView->GetUiaProvider();
+  auto pRootProvider = static_cast<CompositionRootAutomationProvider *>(provider.as<IRawElementProviderSimple>().get());
+  if (pRootProvider != nullptr) {
+    pRootProvider->SetHwnd(m_hwnd);
+  }
+  return provider;
 }
 
 } // namespace winrt::Microsoft::ReactNative::implementation