+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionRootAutomationProvider.cpp
@@ -1,20 +1,18 @@
 #include "pch.h"
 #include "CompositionRootAutomationProvider.h"
-
-#include "Fabric/Composition/CompositionRootView.h"
+#include "UiaHelpers.h"
 
 namespace winrt::Microsoft::ReactNative::implementation {
 
 CompositionRootAutomationProvider::CompositionRootAutomationProvider(
-    winrt::weak_ref<winrt::Microsoft::ReactNative::implementation::CompositionRootView> &&weakRootControl,
-    HWND hwnd) noexcept
-    : m_weakRootControl{std::move(weakRootControl)}, m_hwnd{hwnd} {}
+    const std::shared_ptr<::Microsoft::ReactNative::RootComponentView> &componentView) noexcept
+    : m_view{componentView} {}
 
 // Implementations should return NULL for a top-level element that is hosted in a window. Other elements should return
 // an array that contains UiaAppendRuntimeId (defined in Uiautomationcoreapi.h), followed by a value that is unique
 // within an instance of the fragment.
 //
-// We'll use the View pointer as our identifier for those situations
+// We'll use the react tag as our identifier for those situations
 HRESULT __stdcall CompositionRootAutomationProvider::GetRuntimeId(SAFEARRAY **pRetVal) {
   if (pRetVal == nullptr)
     return E_POINTER;
@@ -108,7 +106,27 @@ HRESULT __stdcall CompositionRootAutomationProvider::GetFocus(IRawElementProvide
 
   *pRetVal = nullptr;
 
+  auto strongView = m_view.view();
+
+  if (strongView == nullptr) {
+    return UIA_E_ELEMENTNOTAVAILABLE;
+  }
+
+  auto rootCV = std::static_pointer_cast<::Microsoft::ReactNative::RootComponentView>(strongView);
+  auto focusedComponent = rootCV->GetFocusedComponent();
+
+  if (focusedComponent) {
+    auto focusedUiaProvider = focusedComponent->EnsureUiaProvider();
+    winrt::com_ptr<IRawElementProviderFragment> spFragment;
+    focusedUiaProvider.as(spFragment);
+    *pRetVal = spFragment.detach();
+  }
+
   return S_OK;
+}
+
+void CompositionRootAutomationProvider::SetHwnd(HWND hwnd) noexcept {
+  m_hwnd = hwnd;
 }
 
 HRESULT __stdcall CompositionRootAutomationProvider::Navigate(
@@ -117,8 +135,12 @@ HRESULT __stdcall CompositionRootAutomationProvider::Navigate(
   if (pRetVal == nullptr)
     return E_POINTER;
 
+  // Fragment roots do not enable navigation to a parent or siblings; navigation among fragment roots is handled by the
+  // default window providers. Elements in fragments must navigate only to other elements within that fragment.
+  if (direction == NavigateDirection_FirstChild || direction == NavigateDirection_LastChild) {
+    return UiaNavigateHelper(m_view, direction, *pRetVal);
+  }
   *pRetVal = nullptr;
-
   return S_OK;
 }