+++ vnext/Microsoft.ReactNative/Fabric/Composition/ImageComponentView.h
@@ -29,10 +29,11 @@ struct WindowsImageResponseObserver;
 
 struct ImageComponentView : CompositionBaseComponentView {
   using Super = CompositionBaseComponentView;
-  ImageComponentView(
+
+  [[nodiscard]] static std::shared_ptr<ImageComponentView> Create(
       const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
       facebook::react::Tag tag,
-      winrt::Microsoft::ReactNative::ReactContext const &reactContext);
+      winrt::Microsoft::ReactNative::ReactContext const &reactContext) noexcept;
 
   void mountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
   void unmountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
@@ -53,6 +54,11 @@ struct ImageComponentView : CompositionBaseComponentView {
   bool focusable() const noexcept override;
 
  private:
+  ImageComponentView(
+      const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
+      facebook::react::Tag tag,
+      winrt::Microsoft::ReactNative::ReactContext const &reactContext);
+
   struct WindowsImageResponseObserver : facebook::react::ImageResponseObserver {
    public:
     WindowsImageResponseObserver(std::shared_ptr<ImageComponentView> image);