+++ vnext/Microsoft.ReactNative/Fabric/Composition/ScrollViewComponentView.h
@@ -51,9 +51,10 @@ struct ScrollInteractionTrackerOwner : public winrt::implements<
 */
 
   using Super = CompositionBaseComponentView;
-  ScrollViewComponentView(
+
+  [[nodiscard]] static std::shared_ptr<ScrollViewComponentView> Create(
       const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
-      facebook::react::Tag tag);
+      facebook::react::Tag tag) noexcept;
 
   void mountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
   void unmountChildComponentView(IComponentView &childComponentView, uint32_t index) noexcept override;
@@ -77,6 +78,10 @@ struct ScrollInteractionTrackerOwner : public winrt::implements<
   bool ScrollWheel(facebook::react::Point pt, int32_t delta) noexcept override;
 
  private:
+  ScrollViewComponentView(
+      const winrt::Microsoft::ReactNative::Composition::ICompositionContext &compContext,
+      facebook::react::Tag tag);
+
   void ensureVisual() noexcept;
   void updateContentVisualSize() noexcept;