+++ vnext/Microsoft.ReactNative/Fabric/ReactTaggedView.h
@@ -15,21 +15,20 @@ namespace Microsoft::ReactNative {
  * changed from the one provided at initalization time (i.e. recycled).
  */
 struct ReactTaggedView {
-  ReactTaggedView(std::shared_ptr<IComponentView> componentView) {
-    m_view = std::move(componentView);
-    m_tag = m_view->tag();
-  }
+  ReactTaggedView(const std::shared_ptr<IComponentView> &componentView)
+      : m_view(componentView), m_tag(componentView->tag()) {}
 
-  std::shared_ptr<IComponentView> &view() noexcept {
-    if (m_view && m_view->tag() != m_tag) {
-      m_view = nullptr;
+  std::shared_ptr<IComponentView> view() noexcept {
+    auto strongView = m_view.lock();
+    if (!m_view.expired() && strongView->tag() != m_tag) {
+      m_view.reset();
     }
-    return m_view;
+    return strongView;
   }
 
  private:
   facebook::react::Tag m_tag;
-  std::shared_ptr<IComponentView> m_view;
+  std::weak_ptr<IComponentView> m_view;
 };
 
 } // namespace Microsoft::ReactNative