+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @react-native-windows/automation-channel
 
-This log was last generated on Wed, 29 Mar 2023 05:13:53 GMT and should not be manually modified.
+This log was last generated on Fri, 31 Mar 2023 05:13:53 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.10.4
+
+Fri, 31 Mar 2023 05:13:53 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.635
 
 ## 0.10.3